// HealthCare/src/Interfaces/ClinicMaster/clinic-response.ts
export interface ClinicResponse {
  clinicId: number;
  clinicCode: string;
  clinicName: string;
  clinicIncharge: string;
  emailId: string;
  mobileNumber: string;
  alternateContactNo: string;
  clinicDoctorCode: string;
  centerCode: string;
  clinicAddress: string;
  rateType: string;
  clinicStatus: boolean;
  createdOn: Date;
  updatedOn: Date;
  createdBy: string;
  updatedBy: string;
  partnerId: string;
}